backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.7
python-decouple==3.8
Pillow==10.0.1
reportlab==4.0.4
celery==5.3.4
redis==5.0.1
django-extensions==3.2.3
python-dateutil==2.8.2

# File parsing dependencies
PyPDF2==3.0.1
pdfplumber==0.9.0
openpyxl==3.1.2
pandas==2.1.3
pytesseract==0.3.10
python-docx==1.1.0